Demand after customer request
Demand after customer request - We added a main screen to each user:
1. Add boxes to see the current state of the number of open / closed tickets, reminder tasks.
2. Adding user information to the front row (all details are taken from the fire base) Picture, alerts and displaying personal information by user - the username and branch where it is located.
3. created TO-DO-LIST for each user. The TO-DO-LIST will be divided into three lists: DO, DOING, and DONE when each task on each list sends to fireBase - and is stored in a separate node for each user tested by its userKey.
For each task that enters TO-DO-LIST, there will be an option to change the status of the task, For example: changing status from DOING to DONE, Changing task urgency - from URGENT to DEFAULT or NEW.

In addition, a change in the opening work card - gave the user an option to narrow the view, if it closes it will only see the title and then the screen will narrow. (For example, there is a part of a car details update that contains fields for adding information such as car description, odometer, engine volume and more - if the user clicks on car details update he will close all fields and display only the title.