Update on the User Management page that available to a admin user only
Update on the User Management page that available to a user with admin privileges only:

Now admin can add new users in a different look than it was.
I added table view by branch of all the active users that are verified in the fireBase.
From each row in the table can be done:

1. Delete an existing user.
2. As part of client requirements - admin can now add a new task to any user he choose from the table.