Add a model to view images and documents that uploaded for each work card:
Add a model to view images and documents that uploaded for each work card:
Get updated photos and documents from the FireBase at each login and after uploading a photo or document.
1. View images that uploaded separately.
2. View documents that uploaded separately.

For each document and image that saved on the server (FireBase) - user can delete and download the document/image .

User can also select multiple photos or multiple documents to download and delete directly!