Changes after customer requirements:
Changes after customer requirements:

Changes in work card page (Update tiacket status):
There was a separate button for uploading photos and a separate button for uploading documents:
The client requested that the upload be shared for both photos and documents:
So I added in the model for photos and documents - a uploading option - now the any kind of file can be uploaded in the same place and the user can upload multiple files at once  - and by the file type there is a separation for photos / documents.
In addition, the client requested to add a progress bar for the progress of each upload of a particular file and to know how many files were loaded.