changes in reports and notification
Changes in Notification: An option to delete all messages or a specific message as add, when no messages are present at all, there is a message that no messages exist.
Changes in daily reports.
And adding a monthly reports page.
We also added an option in daily and monthly reports to produce an Excel report with all the data